+++ module/actor.js
@@ -15,6 +15,7 @@ import { ResourceTrackerManager } from '../module/actor/resource-tracker-manager
 import ApplyDamageDialog from './damage/applydamage.js'
 import * as HitLocations from '../module/hitlocation/hitlocation.js'
 import * as settings from '../lib/miscellaneous-settings.js'
+import { SemanticVersion } from '../lib/semver.js'
 
 export class GurpsActor extends Actor {
   /** @override */
@@ -49,27 +50,22 @@ export class GurpsActor extends Actor {
 
   // Initialize the attribute current values/levels.   The code is expecting 'value' or 'level' for many things, and instead of changing all of the GUIs and OTF logic
   // we are just going to switch the rug out from underneath.   "Import" data will be in the 'import' key and then we will calculate value/level when the actor is loaded.
-  // If import keys don't exist, set them to the current value and commit to upgrade older actors
   _initCurrents() {
+    let v = this.data.data.migrationversion
+    if (!v) return // currently, only need to check for the initial version, but in the future, we might need to check against SemanticVersion.fromString(v)
     // Attributes need to have 'value' set because Foundry expects objs with value and max to be attributes (so we can't use currentvalue)
     let commit = {}
     for (const attr in this.data.data.attributes) {
-      if (this.data.data.attributes[attr].import == null)
-        commit = { ...commit, ...{ ['data.attributes.' + attr + '.import']: this.data.data.attributes[attr].value } }
-      // backward compat
-      else this.data.data.attributes[attr].value = this.data.data.attributes[attr].import
+      this.data.data.attributes[attr].value = this.data.data.attributes[attr].import
     }
     recurselist(this.data.data.skills, (e, k, d) => {
-      if (e.import == null) commit = { ...commit, ...{ ['data.skills.' + k + '.import']: e.level } }
-      else e.level = parseInt(e.import)
+      e.level = parseInt(e.import)
     })
     recurselist(this.data.data.spells, (e, k, d) => {
-      if (e.import == null) commit = { ...commit, ...{ ['data.spells.' + k + '.import']: e.level } }
-      else e.level = parseInt(e.import)
+      e.level = parseInt(e.import)
     })
     recurselist(this.data.data.melee, (e, k, d) => {
-      if (e.import == null) commit = { ...commit, ...{ ['data.melee.' + k + '.import']: e.level } }
-      else e.level = parseInt(e.import)
+      e.level = parseInt(e.import)
       if (!isNaN(parseInt(e.parry))) {
         // allows for '14f' and 'no'
         let base = 3 + Math.floor(e.level / 2)
@@ -87,11 +83,8 @@ export class GurpsActor extends Actor {
       }
     })
     recurselist(this.data.data.ranged, (e, k, d) => {
-      if (e.import == null) commit = { ...commit, ...{ ['data.ranged.' + k + '.import']: e.level } }
-      else e.level = parseInt(e.import)
+       e.level = parseInt(e.import)
     })
-    // We must delay the upgrade of older actor's 'import' keys, since upon startup, the actor may not know which collection it belongs to
-    if (Object.keys(commit).length > 0) setTimeout(() => this.update(commit), 1000)
   }
 
   _applyItemBonuses() {
@@ -372,7 +365,7 @@ export class GurpsActor extends Actor {
     console.log("Importing '" + nm + "'")
     // this is how you have to update the domain object so that it is synchronized.
 
-    let commit = {}
+    let commit = { "data.migrationversion" : game.system.data.version }
 
     if (!game.settings.get(settings.SYSTEM_NAME, settings.SETTING_IGNORE_IMPORT_NAME)) {
       commit = { ...commit, ...{ name: nm } }
@@ -2033,8 +2026,8 @@ export class Named {
 }
 
 export class NamedCost extends Named {
-  constructor() {
-    super()
+  constructor(n1) {
+    super(n1)
     this.points = 0
   }
 }
@@ -2047,16 +2040,16 @@ export class Leveled extends NamedCost {
 }
 
 export class Skill extends Leveled {
-  constructor() {
-    super()
+  constructor(n1, lvl) {
+    super(n1, lvl)
     this.type = '' // "DX/E";
     this.relativelevel = '' // "DX+1";
   }
 }
 
 export class Spell extends Leveled {
-  constructor() {
-    super()
+  constructor(n1, lvl) {
+    super(n1, lvl)
     this.class = ''
     this.college = ''
     this.cost = ''
@@ -2069,8 +2062,8 @@ export class Spell extends Leveled {
 }
 
 export class Advantage extends NamedCost {
-  constructor() {
-    super()
+  constructor(n1) {
+    super(n1)
     this.userdesc = ''
     this.note = ''
   }
@@ -2088,8 +2081,8 @@ export class Attack extends Named {
 }
 
 export class Melee extends Attack {
-  constructor() {
-    super()
+  constructor(n1, lvl, dmg) {
+    super(n1, lvl, dmg)
 
     this.weight = ''
     this.techlevel = ''
@@ -2101,8 +2094,8 @@ export class Melee extends Attack {
 }
 
 export class Ranged extends Attack {
-  constructor() {
-    super()
+  constructor(n1, lvl, dmg) {
+    super(n1, lvl, dmg)
 
     this.bulk = ''
     this.legalityclass = ''